Move scoped_cursor and scoped_session in their own file
