+++ test/cppsuite/CMakeLists.txt
@@ -28,7 +28,8 @@ add_library(cppsuite_test_harness STATIC
     src/main/throttle.cpp
     src/main/transaction.cpp
     src/storage/connection_manager.cpp
-    src/storage/scoped_types.cpp
+    src/storage/scoped_cursor.cpp
+    src/storage/scoped_session.cpp
 )
 
 target_include_directories(cppsuite_test_harness PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
