+++ test/cppsuite/src/component/statistics/statistics.h
@@ -32,7 +32,7 @@
 #include <string>
 
 #include "src/main/configuration.h"
-#include "src/storage/scoped_types.h"
+#include "src/storage/scoped_cursor.h"
 
 namespace test_harness {
